Implement throttling in addition to debouncing in the decorator, rename debounce.py to decorators.py
